[WIP] runbot: self assign builds in a single query
Build self assignation is made in a loop, one at a time.
The row is locked with a select and then the build is scheduled.
From times to times, a psycopg2.extensions.TransactionRollbackError
exception is raised beacause of a parallel update.

With this commit, all available build slots are reserved in a single
LOCKED SQL query.

Unfortunately in very rare cases, the exception still appears.
WIP WIP WIP