[REF] runbot: do not increment next_call with cron_for_host
When cron_for_host is called on the wrong host, the cron job is skipped
but the next_call is incremented of 10 seconds. With a certain amount of
runbot instances, the delay between cron_for_host executions can reach
several minutes. This almost nullify the gain of cron split as the
runbot instances still fight for their cron execution.

With this commit, when the cron_for_host instance raises a custom
exception named CronHostError. This exception is catched by the
overriden callback executing the job and then reraised.
With the help of an ugly monkey patch, this execption is catched by the
cron method that should increase the next_call field. The change on that
field is rollbacked, that way, the cron job can be executed on the
specific host without too much delay.